Add modules method to maintain relationships
