+++ lib/item.rb
@@ -1,33 +1,13 @@
 require 'date'
 
 class Item
-  attr_reader :id, :genre, :author, :source, :label
+  attr_reader :id
   attr_accessor :publish_date
 
   def initialize(publish_date)
     generate_id
     @publish_date = publish_date
     @archived = false
-  end
-
-  def genre=(genre)
-    @genre = genre
-    genre.add_items(self) unless genre.add_items.include?(self)
-  end
-
-  def author=(author)
-    @author = author
-    author.items(self) unless author.items.include?(self)
-  end
-
-  def source=(source)
-    @source = source
-    source.add_items(self) unless source.add_items.include?(self)
-  end
-
-  def label=(label)
-    @label = label
-    label.add_items(self) unless label.add_items.include?(self)
   end
 
   def move_to_archive
